refactor: move geometry related files to new `src/geometry` directory
