+++ src/entities.c
@@ -1,7 +1,7 @@
-#include "collider.h"
 #include "components.h"
 #include "context.h"
 #include "entities.h"
+#include "geometry/collider.h"
 #include "scene.h"
 #include <assert.h>
 #include <raymath.h>
