+++ src/geometry/collider.c
@@ -1,5 +1,5 @@
+#include "../common.h"
 #include "collider.h"
-#include "common.h"
 #include <raymath.h>
 
 Vector2 RectangleRectangleResolution(const Rectangle a, const Rectangle b)
