+++ src/geometry/polygon.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "common.h"
+#include "../common.h"
 #include "line_segment.h"
 
 typedef struct
